change structure for service

+++ src/main/java/ro/lavinia/dto/EmployeeDto.java
@@ -24,7 +24,7 @@ public class EmployeeDto {
 
     private String address;
 
-    private Float salary;
+    private Integer salary;
 
     @JsonIgnore
     private Department department;
